+++ src/refactorings/extract/extract-variable/extract-variable.ts
@@ -181,6 +181,7 @@ function isExtractable(path: t.NodePath): boolean {
     !t.isBlockStatement(path) &&
     !t.isSpreadElement(path) &&
     !t.isTSTypeAnnotation(path) &&
+    !t.isTSTypeParameterInstantiation(path) &&
     // Don't extract object method because we don't handle `this`.
     !t.isObjectMethod(path)
   );